refactoring code, modularizing the components
